+++ addon/routes/list-form.js
@@ -197,6 +197,7 @@ ErrorableRouteMixin, {
           filter: params.filter,
           filterCondition: controller.get('filterCondition'),
           filters: filtersPredicate,
+          filterProjectionName: controller.get('filterProjectionName'),
           predicate: limitPredicate,
           advLimit: advLimit,
           hierarchicalAttribute: hierarchicalAttribute,