+++ tests/dummy/app/router.js
@@ -187,6 +187,7 @@ Router.map(function() {
 
   // Components acceptance tests forms.
   this.route('components-acceptance-tests/flexberry-lookup/settings-example-autofill-by-limit');
+  this.route('components-acceptance-tests/flexberry-lookup/settings-example-custom-window');
   this.route('components-acceptance-tests/flexberry-lookup/base-operations');
   this.route('components-acceptance-tests/flexberry-lookup/settings-example');
   this.route('components-acceptance-tests/flexberry-lookup/settings-example-autocomplete');
@@ -198,6 +199,8 @@ Router.map(function() {
   this.route('components-acceptance-tests/flexberry-lookup/settings-example-preview');
   this.route('components-acceptance-tests/flexberry-lookup/settings-example-preview-page',
   { path: 'components-acceptance-tests/flexberry-lookup/settings-example-preview-page/:id' });
+
+  this.route('components-acceptance-tests/flexberry-simpledatetime/manual-enter');
 
   this.route('components-acceptance-tests/flexberry-groupedit/delete-with-details');
   this.route('components-acceptance-tests/flexberry-objectlistview/base-operations');